utils: Implement a DelayedReader and have the AES CTR deryption use it
Implement a DelayedReader that allows one to wrap an existing io.Reader
and report any errors that that io.Reader may report (at the end)
early to the client while holding back some data from the client.

This patch addresses the issue where the AES decryption has returned
all decrypted data to the client and upon the next Read() from the
cipherstream receives the EOF and 0 bytes but then detects that the HMAC
is not the expected one. In this case the next layer dealing with the
decrypted data would have all the data and could continue. With the
DelayedReader the next layer will run into an error since data were
held back from it.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>