Add pkcs11 support
Extend the parse_helpers with pkcs11 support.

We introduce a configuration file from which pkcs11 parameters are read. A config file
may look like this:

pkcs11:
  module-directories:
    - /usr/lib64/pkcs11/
  allowed-module-paths:
    - /usr/lib64/pkcs11/libsofthsm2.so

The module-directories enumerates directories in which to look for pkcs11 modules.
The allowed-module-paths restricts the pkcs11 modules that can be used.

The config file is searched for in this order:
- ${IMGCRYPT_CONF}
- ${XDG_CONFIG_HOME}/imgcrypt.conf
- ${HOME}/.config/imgcrypt.conf
- /etc/imgcrypt.conf

If no config file can be found then an internal default configuration is used.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>